1.0.0.5 细节
1.0.0.5 细节

+++ rx_orm/rx/rx_entity/rx_table_entity.cs
@@ -8,7 +8,7 @@ namespace rx
 {
     public class rx_table_entity
     {
-        public int row_count { get; set; }
+        public long row_count { get; set; }
 
         public List<rx_entity> rows { get; set; }
     }
